Adds lots of database stuff. Added loadCharacter to EntityFactory which will read a character from the database.

+++ src/main/scala/ayai/apps/GameLoop.scala
@@ -38,8 +38,10 @@ import scala.io.Source
 
 import crane.{EntityProcessingSystem}
 
+import ayai.factories._
+
 object GameLoop {
-  var roomHash : HashMap[Int, Entity] = HashMap.empty[Int, Entity]
+  var roomHash : HashMap[Long, Entity] = HashMap.empty[Long, Entity]
 
   var running : Boolean = _
