+++ src/main/scala/ayai/components/MapChange.scala
@@ -3,6 +3,6 @@ package ayai.components
 import crane.Component
 //this is a temp class, will move this type of messaging into network system
 
-class MapChange(var roomId : Int) extends Component {
+class MapChange(var roomId : Long) extends Component {
 	
 }
