+++ src/main/scala/ayai/components/Room.scala
@@ -2,6 +2,6 @@ package ayai.components
 
 import crane.Component
 
-class Room(val id: Int) extends Component {
+class Room(val id: Long) extends Component {
   
 }
